Restrict pixel size to be a integer multiple of 10
Sometimes, when printing the bars, some bars were visually different from each other in height, even if the value of the pixel, got from the property height, was equal. This bug is probably in the browser side (Opera) which cannot render those pixels with accuracy. At the beggining, the values were float, and after that was a integer, and after that multiples of some number. Multiples of 10 seems about right.